+++ src/br/com/zup/almoco/Almoco.java
@@ -1,0 +1,16 @@
+package br.com.zup.almoco;
+
+import br.com.zup.refeicoesDiarias.RefeicoesDiarias;
+
+public class Almoco extends RefeicoesDiarias {
+
+    public Almoco(String nomeDoAlimento, double qtdDeCaloria) {
+        super(nomeDoAlimento, qtdDeCaloria);
+    }
+
+    @Override
+    public String toString() {
+        return super.toString();
+    }
+}
+
