+++ src/br/com/zup/cafeDaManha/CafeDaManha.java
@@ -1,0 +1,18 @@
+package br.com.zup.cafeDaManha;
+
+import br.com.zup.refeicoesDiarias.RefeicoesDiarias;
+
+public class CafeDaManha extends RefeicoesDiarias {
+
+    public CafeDaManha() {
+    }
+
+    public CafeDaManha(String nomeDoAlimento, double qtdDeCaloria) {
+        super(nomeDoAlimento, qtdDeCaloria);
+    }
+
+    @Override
+    public String toString() {
+        return super.toString();
+    }
+}
