+++ src/br/com/zup/imc/Imc.java
@@ -20,22 +20,21 @@ public class Imc {
         this.pessoa = pessoa;
     }
 
-    public double calculoImc(){
-        double calculo = Math.round(pessoa.getPeso() / (pessoa.getAltura()* pessoa.getAltura()));
+    public double calculoImc() {
+        double calculo = Math.round(pessoa.getPeso() / (pessoa.getAltura() * pessoa.getAltura()));
         return calculo;
     }
 
     private final double ABAIXO_DO_PESO = 18.5;
     private final double PESO_NORMAL = 24.9;
 
-    public void faixasDePeso(){
-        if(calculoImc() < ABAIXO_DO_PESO){
+    public void faixasDePeso() {
+        if (calculoImc() < ABAIXO_DO_PESO) {
             System.out.println("Abaixo do Peso");
-        }else if(calculoImc() <= PESO_NORMAL){
+        } else if (calculoImc() <= PESO_NORMAL) {
             System.out.println("Dentro do peso ideal");
-        }else {
+        } else {
             System.out.println("SobrePeso");
         }
     }
-
 }
