+++ src/br/com/zup/lancheDaTarde/LancheDaTarde.java
@@ -1,0 +1,16 @@
+package br.com.zup.lancheDaTarde;
+
+public class LancheDaTarde extends br.com.zup.refeicoesDiarias.RefeicoesDiarias {
+
+    public LancheDaTarde() {
+    }
+
+    public LancheDaTarde(String nomeDoAlimento, double qtdDeCaloria) {
+        super(nomeDoAlimento, qtdDeCaloria);
+    }
+
+    @Override
+    public String toString() {
+        return super.toString();
+    }
+}
