+++ src/br/com/zup/pessoa/Pessoa.java
@@ -38,10 +38,10 @@ public class Pessoa {
 
     @Override
     public String toString() {
-       StringBuffer retorno = new StringBuffer();
-       retorno.append("Nome da pessoa: " + nome);
-       retorno.append("\nPeso da pessoa: " + peso);
-       retorno.append("\nAltura da pessoa: " + altura);
-       return retorno.toString();
+        StringBuffer retorno = new StringBuffer();
+        retorno.append("Nome da pessoa: " + nome);
+        retorno.append("\nPeso da pessoa: " + peso);
+        retorno.append("\nAltura da pessoa: " + altura);
+        return retorno.toString();
     }
 }
