+++ src/br/com/zup/sistema/Sistema.java
@@ -1,11 +1,11 @@
-package br.com.zup.Sistema;
+package br.com.zup.sistema;
 
-import br.com.zup.Almoco.Almoco;
-import br.com.zup.CafeDaManha.CafeDaManha;
-import br.com.zup.Cardapio.Cardapio;
+import br.com.zup.almoco.Almoco;
+import br.com.zup.cafeDaManha.CafeDaManha;
+import br.com.zup.cardapio.Cardapio;
 import br.com.zup.Jantar.Jantar;
-import br.com.zup.LancheDaTarde.LancheDaTarde;
-import br.com.zup.RefeicoesDiarias.RefeicoesDiarias;
+import br.com.zup.lancheDaTarde.LancheDaTarde;
+import br.com.zup.refeicoesDiarias.RefeicoesDiarias;
 import br.com.zup.imc.Imc;
 import br.com.zup.pessoa.Pessoa;
 
@@ -24,6 +24,7 @@ public class Sistema {
         System.out.println("Digite [2] para ver cardapio:");
         System.out.println("Digite [3] se deseja sair do Programa:");
     }
+
     //menu secundario
     public static void menu2() {
         System.out.println("Em qual refeição deseja colocar esses alimentos");
@@ -78,10 +79,10 @@ public class Sistema {
     }
 
     public static void run() {
-        Cardapio cardapio = new Cardapio();
-        RefeicoesDiarias refeicoesDiarias = new RefeicoesDiarias();
 
         boolean menu = true;
+        Cardapio cardapio = new Cardapio();
+        RefeicoesDiarias refeicoesDiarias = new RefeicoesDiarias();
 
         System.out.println("------------- BEM VINDOS AO DIETA SAUDAVEL -------------");
         Imc imc = new Imc(informacoesDoPaciente());
@@ -90,7 +91,7 @@ public class Sistema {
         imc.faixasDePeso();
         while (menu) {
             menu();
-            int opcoes = dadosDeUsuario("").nextInt();
+            int opcoes = dadosDeUsuario("Escolha sua opção").nextInt();
 
             if (opcoes == 1) {
                 boolean menu2 = true;
@@ -140,14 +141,15 @@ public class Sistema {
                 }
             } else if (opcoes == 2) {
                 System.out.println(cardapio);
-                System.out.println("Total de calorias do dia: " + refeicoesDiarias.somarCaloria(refeicoesDiarias.getQtdDeCaloria()));
+                System.out.println("Total de calorias do dia: " + refeicoesDiarias.getQtdDeCaloria());
             } else if (opcoes == 3) {
-                System.out.println("Voce saiu do programa!!");
+                System.out.println("Voce saiu do menu!!");
                 menu = false;
             } else {
                 System.out.println("Digite um valor válido!");
             }
         }
+        System.out.println("--- FIM DO PROGRAMA ---");
     }
 }
